requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.19
alembic==1.11.1